requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.5.0
async-timeout==4.0.3
attrs==24.2.0
blinker==1.8.2
boto3==1.35.24
botocore==1.35.24
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
cohere==5.9.4
dataclasses-json==0.6.7
exceptiongroup==1.2.2
faiss-cpu==1.8.0.post1
fastavro==1.9.7
filelock==3.16.1
Flask==3.0.3
Flask-Cors==5.0.0
frozenlist==1.4.1
fsspec==2024.9.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.25.0
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.16
langchain-community==0.2.17
langchain-core==0.2.41
langchain-text-splitters==0.2.4
langsmith==0.1.125
MarkupSafe==2.1.5
marshmallow==3.22.0
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.1
numpy==1.24.4
orjson==3.10.7
packaging==24.1
parameterized==0.9.0
pillow==10.4.0
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
s3transfer==0.10.2
safetensors==0.4.5
scikit-learn==1.3.2
scipy==1.10.1
sentence-transformers==3.1.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
sympy==1.13.3
tenacity==8.5.0
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.4.1
tqdm==4.66.5
transformers==4.44.2
types-requests==2.31.0.6
types-urllib3==1.26.25.14
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==1.26.20
Werkzeug==3.0.4
yarl==1.11.1
zipp==3.20.2